Widen site_id to UInt32 so an instance can hold more than 65535 sites
sites.site_id is a Postgres serial: it counts to 2.1 billion and never
reuses a number, not even after a site is deleted. The ClickHouse tables
declared the same column as UInt16, which stops at 65535.

ClickHouse does not reject an id past that. It stores the value modulo the
column width, so site 70000 is written as 4464 and its pageviews are filed
under whichever site really holds that id, surfacing in another customer's
reports. POST /api/track answers {"success":true}, nothing is logged, and
the site that overflowed shows no data at all. Verified against ClickHouse
26.3.17.4.

Two other paths could show one customer's traffic to another, so they are
closed here as well:

- A deleted site left its rows behind. Deletion cleared only the two replay
  tables, so pageviews, bot events and observations stayed in ClickHouse for
  good - harmless until an id is handed out twice, at which point the new
  owner inherits a stranger's history. Every table carrying site_id is now
  cleared. The list is read from ClickHouse rather than hardcoded, but only
  the MergeTree family: a materialized view also carries site_id and would
  reject the DELETE, failing the cleanup after some tables were already
  cleared. Tables parked by a migration as a rollback target are skipped too.

- The public site id had no unique index. It is six random bytes, and site
  lookup takes the first row it finds, so a collision would quietly file one
  site's events against another's; the odds grow with the square of the site
  count. The column is now unique, and both places that mint one retry with
  fresh bytes instead of failing. Which constraint rejected the insert
  decides that. The migration refuses to add the constraint over existing
  duplicates with a message naming them, because docker-entrypoint.sh runs
  migrations under `set -e` before the server starts: a raw index error there
  turns a rare data problem into a container that will not boot. Rows with no
  public id are not duplicates - a unique constraint treats NULLs as distinct.

site_id leads ORDER BY (site_id, timestamp) and ClickHouse refuses ALTER
MODIFY on a key column, so existing installations need the one-off migration
in ops/migrations: it recreates each table from its own live DDL and swaps it
in with EXCHANGE TABLES, keeping the original as <table>__pre_uint32 and
refusing to overwrite one that is already there. Rows are copied a day at a
time - a single INSERT SELECT does not survive a small host, where ClickHouse
caps its own memory near 1 GiB and already holds most of it.

Alongside it: a purge for rows left by earlier deletions, which spares
anything written in the last hour so a site created mid-run does not lose the
first events its owner is watching for; and a snapshot script that captures
both schema and rows, so a restore can rebuild an instance that has nothing
in it rather than only refill tables that already exist.

A test walks the server sources and fails if either half of the width change
reappears narrower than 32 bits. It keys off the column name in the SQL, not
the parameter's name: this codebase already binds a site id as {site:Int32}
and {newSites:Array(Int32)}, so a guard that looked for "siteId" would have
watched the wrong places.

The ops/migrations scripts were then reviewed adversarially and fixed where
they broke. The migration verified row counts with a plain count(), which a
Replacing, Summing or Aggregating table legitimately fails because the fresh
copy merges as it is written - it refused to swap and left instances half
migrated. It swallowed errors from the query that lists days, so a memory
failure quietly became the one unbounded INSERT the per-day copying exists to
avoid. It sliced only by `timestamp`, missing the tables that use event_hour,
session_hour or start_time. It could not tell a run interrupted between the
swap and the rename from a finished one, and reported nothing to do while the
documented rollback no longer worked. And its suffix filters used LIKE, where
an underscore matches any character, so an ordinary table ending in _u32 was
skipped for good.

The snapshot counted rows by asking the table a second time, which disagrees
with the dump whenever a merge lands between the two; it now counts the dump
file itself. Rybbit's own deployment removes system.query_log, so nothing here
depends on it. Restore checks Postgres against its own manifest rather than
trusting pg_restore silently, refuses a target that is non-empty in either
engine, and validates every file it needs before touching anything.

+++ server/src/api/analytics/getRetention.ts
@@ -36,7 +36,7 @@ WITH UserFirstPeriod AS (
         ${effectiveUserId()} AS effective_user_id,
         ${periodFunction}(min(timestamp)${retentionMode === "week" ? ", 1" : ""}) AS cohort_period
     FROM events
-    WHERE site_id = {siteId:UInt16}
+    WHERE site_id = {siteId:UInt32}
     -- Use the configurable time range
     AND timestamp >= addDays(today(), -{timeRange:UInt16})
     GROUP BY effective_user_id
@@ -46,7 +46,7 @@ PeriodActivity AS (
         ${effectiveUserId()} AS effective_user_id,
         ${periodFunction}(timestamp${retentionMode === "week" ? ", 1" : ""}) AS activity_period
     FROM events
-    WHERE site_id = {siteId:UInt16}
+    WHERE site_id = {siteId:UInt32}
     -- Match the date range filter
     AND timestamp >= addDays(today(), -{timeRange:UInt16})
 ),
